refactor(aces): split modules and cut returns/complexity for SonarCloud gate
No behavior change. Extract placement-realization dispatch into
aces_placement_realization.py and account-parity provisioner parsing into
validation/_account_parity.py so aces_realization.py and _gate_checks.py
drop back under the 500-line limit; reduce the fail-closed resolvers to
<=3 returns via reject-reason helpers (distinct reason codes preserved)
and extract per-attribute helpers to cut function/cognitive complexity.
